Chore: remove readrate limiting
Remove the readrate limiter so players can buffer ahead what they want, otherwise things like crossfade may not work well. We can bring this back later as an optional toggle if we want.

Downside of disabling the readrate limiter is that when you want to schedule a next track to play in the queue, the player may already have buffered ahead.