Fix: Subsonic: Use any parent link when enumerating songs (#1966)
Fix: Subonsic: Use any parent link when enumerating songs

The LMS implementation does not use the parent field for album IDs. The
spec allows for one or both of parent and albumId to be filled so we
should accept both.

Also fix typing info for the local staring the fetched album.

Closes: music-assistant/support#3501

Signed-off-by: Eric B Munson <eric@munsonfam.org>
Co-authored-by: Marcel van der Veldt <m.vanderveldt@outlook.com>