Fix: race condition in provider config handling
The config object within a provider is meant to be read-only but it was being used by providers also to update refresh tokens etc. This lead to race conditions because the provider has a cached copy of the config object. Also a failed login when the provider was already loaded was not properly handled. That is all adjusted now.